refactor(warehouse): rename module to index
shorter import paths in other modules
