+++ src/syntax/tree/index.ts
@@ -20,7 +20,7 @@ import {
     TreeNodeStatement,
     TreeNodeBlock,
 } from './node';
-import { addInstance, getInstance, removeInstance } from '../warehouse/warehouse';
+import { addInstance, getInstance, removeInstance } from '../warehouse';
 import { checkValueAssignment, queryElementSpecification } from '../specification/specification';
 
 import { ElementArgument } from '../elements/elementArgument';
