+++ src/syntax/warehouse/index.ts
